共通暗号ユーティリティの追加と既存コードの整理
