+++ shared/crypto.ts
@@ -1,0 +1,35 @@
+export async function sha256Hex(text: string): Promise<string> {
+  const buf = new TextEncoder().encode(text);
+  const hash = await crypto.subtle.digest("SHA-256", buf);
+  return Array.from(new Uint8Array(hash))
+    .map((b) => b.toString(16).padStart(2, "0"))
+    .join("");
+}
+
+export function bufferToPem(
+  buffer: ArrayBuffer,
+  type: "PRIVATE KEY" | "PUBLIC KEY",
+) {
+  const b64 = btoa(String.fromCharCode(...new Uint8Array(buffer)));
+  const lines = b64.match(/.{1,64}/g)?.join("\n") ?? b64;
+  return `-----BEGIN ${type}-----\n${lines}\n-----END ${type}-----`;
+}
+
+export async function generateKeyPair() {
+  const pair = await crypto.subtle.generateKey(
+    {
+      name: "RSA-PSS",
+      modulusLength: 2048,
+      publicExponent: new Uint8Array([1, 0, 1]),
+      hash: "SHA-256",
+    },
+    true,
+    ["sign", "verify"],
+  );
+  const priv = await crypto.subtle.exportKey("pkcs8", pair.privateKey);
+  const pub = await crypto.subtle.exportKey("spki", pair.publicKey);
+  return {
+    privateKey: bufferToPem(priv, "PRIVATE KEY"),
+    publicKey: bufferToPem(pub, "PUBLIC KEY"),
+  };
+}
